+++ packages/ui-test/src/component/Test.svelte
@@ -1,0 +1,6 @@
+<script lang='ts'>
+  export let variable: number
+</script>
+
+Test component
+{variable}