+++ packages/ui-test/src/index.ts
@@ -1,0 +1,4 @@
+import Test from './component/Test.svelte'
+export {
+  Test
+}