+++ packages/platform-rig/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @hcengineering/platform-rig
 
-This log was last generated on Sun, 12 Oct 2025 02:49:17 GMT and should not be manually modified.
+This log was last generated on Sun, 12 Oct 2025 03:31:47 GMT and should not be manually modified.
+
+## 0.7.17
+Sun, 12 Oct 2025 03:31:47 GMT
+
+### Patches
+
+- Fix svelte build
 
 ## 0.7.16
 Sun, 12 Oct 2025 02:49:17 GMT